refactor: toggle screen, dropdown menu and canvas
