+++ packages/react-router-v7-example/app/routes.ts
@@ -1,3 +1,3 @@
 import { type RouteConfig, index } from '@react-router/dev/routes';
 
-export default [index('routes/index.tsx')] satisfies RouteConfig;
+export default [index('routes/index/route.tsx')] satisfies RouteConfig;
